+++ NEWS.rst
@@ -5,6 +5,7 @@ News
 Version 0.8.9 - dev
 
    - Release notes: https://ezdxf.mozman.at/release-v0-8-9.html
+   - NEW: Drawing.layouts_and_blocks(), iterate over all layouts (mode space and paper space) and all block definitions.
    - `Basic` read support for all missing DXF entities/objects:
        - ACAD_PROXY_GRAPHIC
        - HELIX